move wrapper from init
